\encoding{utf-8}
\name{auta2012}
\alias{auta2012}
\alias{auta2012_eng}
\docType{data}
\title{Offer Prices of Second-hand Cars in 2012}
\description{
The data set auta2012 was designed with a view of creating opportunity to work on a big and interesting data set concerning transactions. Transaction data is usually rather extensive and it is processed in a very specific manner. 

This data was extracted from the announcement service otomoto.pl in 2012. The set contains information about over 207 thousand car sales offers. 

Each offer (each row) is described by 21 variables including:

- price, currency and information whether the given price is a net or a gross price,

-	information on engine power expressed in horsepower and kW and its capacity,

-	information about model, make and version of the car,

-	information on the type of drive, year of manufacture and mileage (in km),

-	information about color, number of doors, country of origin, country of registration and accessories.

Information about accessories is described by a very long line of text with values separated by comas.

[POL]

Ceny ofertowe używanych aut w roku 2012

Zbiór danych auta2012 przygotowaliśmy po to, by dać możliwość pracy z dużym i ciekawym zbiorem danych o transakcjach. Dane transakcyjne są zazwyczaj bardzo długie, a sposób ich przetwarzania jest specyficzny.

Dane zostały pobrane w roku 2012 z serwisu ogłoszeń otomoto.pl. Zbiór danych zawiera informacje o ofertach dla przeszło 207 tysięcy ogłoszeń sprzedaży auta.

Każda oferta (każdy wiersz) opisana jest przez 21 zmiennych, w tym:

- cenę, walutę i informację czy jest to cena brutto czy netto,

- informację o mocy silnika w koniach mechanicznych i kW, oraz jego pojemności,

- informację o modelu, marce, wersji auta,

- informację o rodzaju napędu, roku produkcji oraz przebiegu (w km),

- informację o kolorze, liczbie drzwi, kraju pochodzenia, kraju rejestracji oraz wyposażeniu.

Informacja o wyposażeniu jest opisana przez napis z wartościami rozdzielanymi przecinkiem, przez co sam napis może być bardzo długi.
}
\author{
 Przemyslaw Biecek
}
\examples{
head(auta2012)
}
\keyword{datasets}
