\name{Diagnoza-Spoleczna}
\alias{diagnoza}
\alias{diagnozaDict}
\docType{data}
\title{A Subset of Polish Social Diagnosis Data}
\description{
This data is a subset of Diagnoza dataset available on github (https://github.com/pbiecek/Diagnoza).

Description of consecutive columns is in the diagnozaDict object.

Source: http://diagnoza.com/
Full dataset: https://github.com/pbiecek/Diagnoza
}
\author{
Przemyslaw Biecek
}

