\encoding{utf-8}
\name{galton}
\alias{galton}
\alias{pearson}
\docType{data}
\title{Galton's and Pearson's Height Data for Parents and Children}
\description{
Height of parents and children

The word regression comes from the pioneering studies carried out by Francis Galton and Karl Pearson concerning relationship between height of parents and their children’s. 

The original data sets created by the scientists are available in the variables galton and pearson. Using these data sets we can describe relationship between the height of a son and his father (pearson) and relationship between the weighted average height of the parents (galton) and their son’s height. 

This is a data set used by Galton in 1885 to study the relationship between a parent's height and their childrens.
The second dataset was used by Pearson to investigate regression. 

Inches were converted into centimeters.

There are 928 rows in the galton data set and 1078 rows in the pearson data set. 

[POL]

Wzrost rodziców i dzieci

Słowo regresja wywodzi się z pionierskich badań Francisa Galtona i Karla Pearsona nad zależnością wzrostu dzieci i rodziców.

Oryginalne zbiory danych obu tych badaczy są dostępne w zmiennych galton i pearson. Na bazie tych zbiorów danych można opisywać zależność pomiędzy wzrostem syna a ojca (pearson) oraz zależność pomiędzy ważoną średnią z wzrostu rodziców (galton) z wzrostem syna.

Poniżej przedstawiamy pierwszych 6 wierszy z każdego ze zbiorów danych. W zbiorze danych galton znajduje się 928 wierszy a w zbiorze danych pearson znajduje się 1078 wierszy.

}
\author{
 Copy from other packages with inches transformed into centimeters.
}
\examples{
head(galton)
head(pearson)
}
\keyword{datasets}
