\encoding{utf-8}
\name{koty_ptaki}
\alias{koty_ptaki}
\alias{cats_birds}
\docType{data}
\title{The Fastest Cats and Birds in the World}
\description{
Size, habitat, speed and weight of big cats

The data set cats_birds was created as a so-called toy-example. It is a small set (13 rows and 7 columns) which can be fully displayed on the computer screen. You can use it to practice basic operations on the data. 

The data set juxtaposes characteristics of 13 chosen species of cats and birds presented in separate rows. The chosen species are the fastest members of both groups. Each row contains information on a different species. 

The columns of this data set present respectively:

- name of the species,

-	maximum weight of an individual animal expressed in kilograms and maximum body length in meters, 

-	maximum speed (in the horizontal flight in case of birds, not during the dive) expressed in kilometers per hour,

-	habitat and length of life in years,

-	the last column specifies whether the species is a big cat or a bird.


[POL]

Najszybsze koty i ptaki na świecie

Zbiór danych koty_ptaki powstał jako tzw. ,,toy-example’’. Jest to niewielki (13 wierszy 7 kolumn) zbiór, który można w całości wyświetlić na ekranie i na którym można przećwiczyć podstawowe operacje na danych.

W danych zestawiono charakterystyki dla 13 wybranych gatunków przedstawionych w kolejnych wierszach. Wybrane gatunki to najszybsi przedstawiciele kotów i ptaków. Każdy wiersz to informacje o innym gatunku.

Kolejne kolumny w tym zbiorze danych przedstawiają:

- nazwę gatunku,

- maksymalną osobniczą wagę w kilogramach i maksymalną długość ciała w metrach,

- maksymalną prędkość (dla ptaków w locie poziomym, nie w nurkowaniu) w kilometrach na godzinę,

- obszar zamieszkania oraz żywotność w latach,

- ostatnia kolumna określa czy gatunek jest dużym kotem czy ptakiem.
}
\author{
 Przemyslaw Biecek
}
\examples{
head(koty_ptaki)
}
\keyword{datasets}
