\encoding{utf-8}
\name{mandatySejmik2014}
\alias{mandatySejmik2014}
\alias{votes2014}
\docType{data}
\title{Local Government Elections in Poland 2014}
\description{

The data set votes2014 / mandatySejmik2014 was prepared to show ways in which spatial data can be analyzed and visualized. The data was downloaded from the website of the National Electoral Commission (http://wybory2014.pkw.gov.pl/pl/). 

The data presents information about the number of seats in the regional parliaments in Poland won in 2014. The set consists of 16 rows and 9 columns. Each row presents results concerning a different voivodeship. 

The subsequent columns present respectively:

- name of the voivodeship,

-	number of seats won by PSL, PiS, PO, SLD and other political parties,

-	what is the percentage of valid votes of the total number of votes cast by the people entitled to vote,

-	longitude and latitude of the geographical centre of the voivodeship.

[POL]

Wyniki wyborów samorządowych w 2014

Zbiór danych mandatySejmik2014 przygotowaliśmy po to, by pokazać jak można analizować lub wizualizować dane przestrzenne. Dane pobrano ze stron Państwowej Komisji Wyborczej (http://wybory2014.pkw.gov.pl/pl/).

Dane przedstawiają informacje o liczbie wygranych mandatów w sejmikach w wyborach samorządowych w Polsce w roku 2014. Zbiór danych zawiera 16 wierszy i 9 kolumn. Każdy wiersz przedstawia wyniki dla innego województwa.

Kolejne kolumny opisują:

- nazwę województwa,

- liczbę zdobytych mandatów przez PSL, PiS, PO, SLD i inne partie,

- jaki procent z uprawnionych do głosowania stanowi procent ważnych głosów,

- długość i szerokość geograficzną środka geograficznego województwa.

}
\author{
 Data from http://wybory2014.pkw.gov.pl/pl/
}
\examples{
head(mandatySejmik2014)
}
\keyword{datasets}
