\encoding{utf-8}
\name{PogromcyDanych-package}
\alias{PogromcyDanych-package}
\alias{PogromcyDanych}
\docType{package}
\title{
R package for the PogromcyDanych Masive Online Open Course
}
\description{
The data sets used in the online course ,,PogromcyDanych''. The course about R, visualization and data analysis. The dataset includes both large data sets, such as auta2012 with prices for cars offered for sale in 2012, as well as small data sets to basic exercises such as koty_ptaki.
}
\details{
\tabular{ll}{
Package: \tab PogromcyDanych\cr
Type: \tab Package\cr
Version: \tab 1.4\cr
Date: \tab 2015-01-05\cr
License: \tab GPL-3 \cr
}
}
\author{
Author: Przemyslaw Biecek (Foundation SmarterPoland.pl)
Maintainer: Przemyslaw Biecek <przemyslaw.biecek@gmail.com>
}
\references{
http://www.PogromcyDanych.pl
}
\keyword{ package }
