\name{serialeIMDB}
\alias{serialeIMDB}
\alias{serialsIMDB}
\docType{data}
\title{Ratings of 200 series from IMDb}
\description{
Ratings of episodes for selected series were downloaded.

For example ratings for Ninjago series are avaliable here
http://www.imdb.com/title/tt1871731
}
\author{
 Przemyslaw Biecek [R convertion]
}
\examples{
head(serialeIMDB)
}
\keyword{datasets}
