\encoding{utf-8}
\name{PogromcyDanych-package}
\alias{PogromcyDanych-package}
\alias{PogromcyDanych}
\alias{DataCrunchers}
\docType{package}
\title{
PogromcyDanych / DataCrunchers is the Massive Online Open Course that Brings R and Statistics to the People
}
\description{
The data sets used in the online course ,,PogromcyDanych''. The package includes both large data sets, such as auta2012 with prices for cars offered for sale in 2012, as well as small data sets for basic exercises such as cats_birds.

You can process data in many ways. The course Data Crunchers will introduce you to this variety. For this reason we will work on datasets of different size (from several to several hundred thousand rows), with various level of complexity (from two to two thousand columns) and prepared in different formats (text data, quantitative data and qualitative data).

All of these data sets were gathered in a single big package called PogromcyDanych to facilitate access to them. It contains all sorts of data sets such as data about offer prices of cars, results of opinion polls, information about changes in stock market indices, data about names given to newborn babies, ski jumping results or information about outcomes of breast cancer patients treatment.

[POL]

Praca z danymi może być bardzo różnorodna. Podczas kursu ,,Pogromcy Danych’’ będziemy tę różnorodność przybliżać. W tym celu będziemy pracować na zbiorach danych o różnej wielkości (od kilkunastu do kilkuset tysięcy wierszy), różnym stopniu złożoności (od dwóch do dwóch tysięcy kolumn), oraz o różnym formacie (dane tekstowe, ilościowe, jakościowe).

Aby ułatwić dostęp do tak różnych zbiorów danych zebraliśmy je wszystkie w jednym pakiecie o nazwie PogromcyDanych. Znaleźć można w nim najróżniejsze dane, w tym o cenach ofertowych aut, dane ankietowe dotyczące badań społecznych, informacje o zmianach indeksów giełdowych, imionach noworodków, wynikach skoków narciarskich czy o wynikach leczenia pacjentek z nowotworem piersi.
}
\author{
Author: Przemyslaw Biecek
Maintainer: Przemyslaw Biecek <przemyslaw.biecek@gmail.com>
}
\references{
http://www.PogromcyDanych.pl
}
\keyword{ package }
