\name{TCGA_BRCA}
\alias{TCGA_BRCA}
\docType{data}
\title{Selected Variables from BReast CAncer Patients of The Cancer Genome Atlas Project}
\description{
The Cancer Genome Atlas study collects clinical and genomic data from different tumors. 
This dataset incorporates selected variables from patients with BReast CAncer.

More details about TCGA project: http://cancergenome.nih.gov/
}
\author{
 The Cancer Genome Atlas (collection), Przemyslaw Biecek (transformation)
}
\examples{
head(TCGA_BRCA)
}
\keyword{datasets}
